mobile_shopping_agent/requirements.txt
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=0.24.2
nltk>=3.6.3
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.0
pytorch>=1.9.0
transformers>=4.11.0
spacy>=3.1.3
python-dateutil>=2.8.2
google-generativeai==0.8.5
pandas
Flask
python-dotenv
numpy
scikit-learn
tensorflow
nltk
Pillow==10.3.0
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'
